+++ app/src/main/java/fr/free/nrw/commons/MediaDataExtractor.java
@@ -26,6 +26,7 @@ import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.media.Media;
 import fr.free.nrw.commons.mwapi.MediaResult;
 import fr.free.nrw.commons.mwapi.MediaWikiApi;
+import fr.free.nrw.commons.utils.Utils;
 import timber.log.Timber;
 
 /**
