+++ app/src/main/java/fr/free/nrw/commons/upload/SingleUploadFragment.java
@@ -48,7 +48,7 @@ import butterknife.OnClick;
 import butterknife.OnItemSelected;
 import butterknife.OnTouch;
 import fr.free.nrw.commons.R;
-import fr.free.nrw.commons.Utils;
+import fr.free.nrw.commons.utils.Utils;
 import fr.free.nrw.commons.di.CommonsDaggerSupportFragment;
 import fr.free.nrw.commons.settings.Prefs;
 import fr.free.nrw.commons.utils.ViewUtil;
